// src/types.ts
export interface StationInformation {
  station_id: string;
  name: string;
  address: string;
  lat: number;
  lon: number;
  capacity: number;
}

export interface StatusStation {
  station_id: string;
  is_installed: number;
  is_renting: number;
  num_bikes_available: number;
  num_docks_available: number;
  last_reported: number;
  is_returning: number;
}

export interface Station extends StationInformation, StatusStation {}